Separate Searchbar into an own component
