+++ src/components/SearchBar/DropdownMenu.js
@@ -1,0 +1,5 @@
+import React, {useState} from "react";
+
+export default function Dropdown({pokemons}) {
+    
+}
